Add pillars to the city and other things

+++ src/generation/pillar.rs
@@ -9,11 +9,11 @@ pub struct Pillar {
   min_height: i32,
   max_height: i32,
   origin: IVec2,
-  radius: i32
+  radius: u32
 }
 
 impl Pillar {
-  pub fn new(origin: IVec2, radius: i32) -> Self {
+  pub fn new(origin: IVec2, radius: u32) -> Self {
     Pillar {
       min_height: crate::WORLD_Z_MIN,
       max_height: crate::WORLD_Z_MAX,
@@ -22,7 +22,7 @@ impl Pillar {
     }
   }
 
-  pub fn new_bounded(origin: IVec2, radius: i32, min: Option<i32>, max: Option<i32>) -> Self {
+  pub fn new_bounded(origin: IVec2, radius: u32, min: Option<i32>, max: Option<i32>) -> Self {
     Pillar {
       min_height: min.unwrap_or(crate::WORLD_Z_MIN),
       max_height: max.unwrap_or(crate::WORLD_Z_MAX),
@@ -34,7 +34,7 @@ impl Pillar {
 
 impl Geometry for Pillar {
   fn bounding_box_guess(&self) -> BoundingBox {
-    let o = IVec2::splat(self.radius + 1);
+    let o = IVec2::splat((self.radius + 1) as i32);
     let min = (self.origin - o).extend(self.min_height);
     let max = (self.origin + o).extend(self.max_height);
     BoundingBox::new(min, max)
